move lexer tests to separate module
