+++ src/parser/lexer.rs
@@ -1,3 +1,6 @@
+#[cfg(test)]
+mod tests;
+
 use super::data::*;
 
 pub fn tokenize(query: &str) -> Result<Vec<Token>> {
@@ -115,178 +118,5 @@ impl<'a> Tokenizer<'a> {
 		self.position += count;
 		self.tail = tail;
 		word
-	}
-}
-
-#[cfg(test)]
-mod tests {
-	use super::*;
-
-	#[test]
-	fn spec() {
-		let cases = [
-			("", Ok(vec![])),
-			(
-				"Ph'nglui fhtagn",
-				Ok(vec![
-					Token::Tag {
-						value: "Ph'nglui",
-						start: 0,
-						end: 8,
-					},
-					Token::Tag {
-						value: "fhtagn",
-						start: 9,
-						end: 15,
-					},
-				]),
-			),
-			(
-				"kek\"no... \"",
-				Err(ParseError::UnpairedQuote { position: 10 }),
-			),
-			(
-				"o\"oops \"\"",
-				Ok(vec![
-					Token::Tag {
-						value: "o\"oops",
-						start: 0,
-						end: 6,
-					},
-					Token::Tag {
-						value: "",
-						start: 7,
-						end: 9,
-					},
-				]),
-			),
-			(
-				"\"\"\"\"\"\"",
-				Ok(vec![
-					Token::Tag {
-						value: "",
-						start: 0,
-						end: 2,
-					},
-					Token::Tag {
-						value: "",
-						start: 2,
-						end: 4,
-					},
-					Token::Tag {
-						value: "",
-						start: 4,
-						end: 6,
-					},
-				]),
-			),
-			(
-				"\"\"\"\"\"\"\"",
-				Err(ParseError::UnpairedQuote { position: 6 }),
-			),
-			(
-				"Опа, кириллица. or else?...",
-				Ok(vec![
-					Token::Tag {
-						value: "Опа,",
-						start: 0,
-						end: 7,
-					},
-					Token::Tag {
-						value: "кириллица.",
-						start: 8,
-						end: 27,
-					},
-					Token::Operator {
-						kind: OperatorKind::Or,
-						position: 28,
-					},
-					Token::Tag {
-						value: "else?...",
-						start: 31,
-						end: 39,
-					},
-				]),
-			),
-			(
-				"or;and;not)or_gibberish   (         ",
-				Ok(vec![
-					Token::Operator {
-						kind: OperatorKind::Or,
-						position: 0,
-					},
-					Token::Tag {
-						value: ";and;not",
-						start: 2,
-						end: 10,
-					},
-					Token::Operator {
-						kind: OperatorKind::CloseParen,
-						position: 10,
-					},
-					Token::Operator {
-						kind: OperatorKind::Or,
-						position: 11,
-					},
-					Token::Tag {
-						value: "_gibberish",
-						start: 13,
-						end: 23,
-					},
-					Token::Operator {
-						kind: OperatorKind::OpenParen,
-						position: 26,
-					},
-				]),
-			),
-			(
-				"and~ wow   oR NOT such~ \"a disappointment, not\" ornot?",
-				Ok(vec![
-					Token::Operator {
-						kind: OperatorKind::And,
-						position: 0,
-					},
-					Token::Operator {
-						kind: OperatorKind::Fuzzy,
-						position: 3,
-					},
-					Token::Tag {
-						value: "wow",
-						start: 5,
-						end: 8,
-					},
-					Token::Operator {
-						kind: OperatorKind::Or,
-						position: 11,
-					},
-					Token::Operator {
-						kind: OperatorKind::Not,
-						position: 14,
-					},
-					Token::Tag {
-						value: "such",
-						start: 18,
-						end: 22,
-					},
-					Token::Operator {
-						kind: OperatorKind::Fuzzy,
-						position: 22,
-					},
-					Token::Tag {
-						value: "a disappointment, not",
-						start: 24,
-						end: 47,
-					},
-					Token::Tag {
-						value: "ornot?",
-						start: 48,
-						end: 54,
-					},
-				]),
-			),
-		];
-		for (query, expected) in cases {
-			assert_eq!(tokenize(query), expected);
-		}
 	}
 }
